:construction: Add more thing ! (enchant and food)

+++ src/fr/maxlego08/items/api/configurations/FoodEffect.java
@@ -1,5 +1,6 @@
 package fr.maxlego08.items.api.configurations;
 
-public record FoodEffect(String type, double probability, int amplifier, int duration, boolean showParticles,
+public record FoodEffect(String type, float probability, int amplifier, int duration, boolean showParticles,
                          boolean showIcon, boolean ambient) {
+
 }
